requirements.txt
streamlit-0.87.0-py2.py3-none-any.whl